# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Approval System API
  version: 1.0.0
  description: 基于当前代码实现整理的接口文档（Spring Boot + Flowable）
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/requests:
    get:
      tags: [Requests]
      summary: 按数据权限查询申请单
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BizRequest'
              example:
                - id: 1
                  businessKey: BR-20260224-0001
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  processDefinitionId: approvalWorkflow:3:7504
                  formInstanceId: 88
                  applicantId: 1001
                  applicantDeptId: 10
                  applicantPostId: 101
                  title: 差旅申请-上海
                  status: 2
                  currentTaskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  currentAssigneeId: 1002
                  submitTime: "2026-02-24T10:20:30Z"
                  finishTime: null
                  createdAt: "2026-02-24T10:20:00Z"
                  updatedAt: "2026-02-24T10:30:00Z"
                  isDeleted: 0
  /api/requests/tasks:
    get:
      tags: [Requests]
      summary: 查询可见流程任务
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskInfo'
              example:
                - taskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  taskName: 部门主管审批
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  assignee: "1002"
                  createTime: "2026-02-24T10:20:30Z"
  /api/requests/logs:
    get:
      tags: [Requests]
      summary: 查询可见审批日志
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BizRequestLog'
              example:
                - id: 101
                  businessKey: BR-20260224-0001
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  taskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  operatorId: 1002
                  action: APPROVE
                  comment: 同意
                  createdAt: "2026-02-24T10:35:00Z"
  /api/requests/processes:
    get:
      tags: [Requests]
      summary: 查询可见流程实例
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessInfo'
              example:
                - processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  processDefinitionId: approvalWorkflow:3:7504
                  businessKey: BR-20260224-0001
...
  /api/rbac/users:
    post:
      tags: [RBAC]
      summary: 创建用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              operatorId: 1
              username: zhangsan
              password: "P@ssw0rd123"
              deptId: 2
              status: 1
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RbacUserResponse'
              example:
                id: 10
                username: zhangsan
                deptId: 2
                status: 1
        '400':
          description: 参数校验失败
        '403':
          description: 权限不足
        '409':
          description: 资源冲突（如用户名重复）
  /api/rbac/roles:
    post:
      tags: [RBAC]
      summary: 创建角色
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
            example:
              operatorId: 1
              roleCode: EMPLOYEE
              roleName: 普通员工
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysRole'
              example:
                id: 3
                roleCode: EMPLOYEE
                roleName: 普通员工
                status: 1
        '400':
          description: 参数校验失败
        '403':
          description: 权限不足
        '409':
          description: 资源冲突（如角色编码重复）
  /api/rbac/depts:
    post:
      tags: [RBAC]
      summary: 创建部门
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeptRequest'
            example:
              operatorId: 1
              deptName: 财务部
              parentId: 1
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDept'
              example:
                id: 2
                parentId: 1
                deptName: 财务部
        '400':
          description: 参数校验失败
        '403':
          description: 权限不足
  /api/rbac/posts:
    post:
      tags: [RBAC]
      summary: 创建岗位
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
            example:
              operatorId: 1
              postCode: FINANCE_MANAGER
              postName: 财务经理
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysPost'
              example:
                id: 5
                postCode: FINANCE_MANAGER
                postName: 财务经理
        '400':
          description: 参数校验失败
        '403':
          description: 权限不足
        '409':
          description: 资源冲突（如岗位编码重复）
  /api/rbac/assign:
    post:
      tags: [RBAC]
      summary: 用户绑定角色
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
            example:
              operatorId: 1
              userId: 10
              roleId: 3
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Role assigned
        '400':
          description: 参数错误（如用户或角色不存在）
        '403':
          description: 权限不足
  /api/rbac/assign-post:
    post:
      tags: [RBAC]
      summary: 用户绑定岗位
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPostRequest'
            example:
              operatorId: 1
              userId: 10
              postId: 5
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Post assigned
        '400':
          description: 参数错误（如用户或岗位不存在）
        '403':
          description: 权限不足
  /api/rbac/role-data-scope:
    post:
      tags: [RBAC]
      summary: 配置角色数据权限
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRoleDataScopeRequest'
            example:
              operatorId: 1
              roleId: 3
              scopeType: DEPT_AND_CHILD
              deptId: 2
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Role data scope added
        '400':
          description: 参数错误（如角色不存在或scopeType非法）
        '403':
          description: 权限不足
components:
  parameters:
...
    UserIdQuery:
      name: userId
      in: query
      required: true
      description: 当前登录用户ID
      schema:
        type: integer
        format: int64
        example: 1001
  schemas:
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        message:
          type: string
          description: 操作结果说明
          example: Task completed
...
    TaskInfo:
      type: object
      properties:
        taskId:
          type: string
          description: 任务ID
          example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
        taskName:
          type: string
          description: 任务名称
          example: 部门主管审批
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        assignee:
          type: string
          nullable: true
          description: 当前处理人
          example: "1002"
        createTime:
          type: string
          format: date-time
          description: 任务创建时间
          example: "2026-02-24T10:20:30Z"
    ProcessInfo:
      type: object
      properties:
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        processDefinitionId:
          type: string
          description: 流程定义ID
          example: approvalWorkflow:3:7504
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
    BizRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 主键ID
          example: 1
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        processDefinitionId:
          type: string
          description: 流程定义ID
          example: approvalWorkflow:3:7504
        formInstanceId:
          type: integer
          format: int64
          nullable: true
          description: 表单实例ID
          example: 88
        applicantId:
          type: integer
          format: int64
          description: 申请人ID
          example: 1001
        applicantDeptId:
          type: integer
          format: int64
          nullable: true
          description: 申请人部门ID
          example: 10
        applicantPostId:
          type: integer
          format: int64
          nullable: true
          description: 申请人岗位ID
          example: 101
        title:
          type: string
          description: 申请标题
          example: 差旅申请-上海
        status:
          type: integer
          description: 0草稿/1已提交/2审批中/3通过/4拒绝/5退回/6撤销
          example: 2
        currentTaskId:
          type: string
          nullable: true
          description: 当前任务ID
          example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
        currentAssigneeId:
          type: integer
          format: int64
          nullable: true
          description: 当前处理人ID
          example: 1002
        submitTime:
          type: string
          format: date-time
          nullable: true
          description: 提交时间
          example: "2026-02-24T10:20:30Z"
        finishTime:
          type: string
          format: date-time
          nullable: true
          description: 完成时间
          example: "2026-02-24T11:20:30Z"
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2026-02-24T10:20:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: "2026-02-24T10:30:00Z"
        isDeleted:
          type: integer
          description: 逻辑删除标记（0否/1是）
          example: 0
    BizRequestLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 主键ID
          example: 100
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        taskId:
          type: string
          nullable: true
          description: 任务ID
          example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
        operatorId:
          type: integer
          format: int64
          description: 操作人ID
          example: 1002
        action:
          type: string
          description: 操作类型
          example: APPROVE
        comment:
          type: string
          nullable: true
          description: 操作说明
          example: 同意
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2026-02-24T10:35:00Z"
...
    CreateUserRequest:
      type: object
      required: [username]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        username:
          type: string
          description: 用户名（唯一）
          example: zhangsan
        password:
          type: string
          nullable: true
          description: 登录密码（明文传输后由后端哈希）
          example: "P@ssw0rd123"
        deptId:
          type: integer
          format: int64
          nullable: true
          description: 部门ID
          example: 2
        status:
          type: integer
          description: 0禁用/1启用
          example: 1
    CreateRoleRequest:
      type: object
      required: [roleCode, roleName]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        roleCode:
          type: string
          description: 角色编码（唯一）
          example: EMPLOYEE
        roleName:
          type: string
          description: 角色名称
          example: 普通员工
    CreateDeptRequest:
      type: object
      required: [deptName]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        deptName:
          type: string
          description: 部门名称
          example: 财务部
        parentId:
          type: integer
          format: int64
          nullable: true
          description: 上级部门ID
          example: 1
    CreatePostRequest:
      type: object
      required: [postCode, postName]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        postCode:
          type: string
          description: 岗位编码（唯一）
          example: FINANCE_MANAGER
        postName:
          type: string
          description: 岗位名称
          example: 财务经理
    AssignRoleRequest:
      type: object
      required: [userId, roleId]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        userId:
          type: integer
          format: int64
          description: 用户ID
          example: 10
        roleId:
          type: integer
          format: int64
          description: 角色ID
          example: 3
    AssignPostRequest:
      type: object
      required: [userId, postId]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        userId:
          type: integer
          format: int64
          description: 用户ID
          example: 10
        postId:
          type: integer
          format: int64
          description: 岗位ID
          example: 5
    AddRoleDataScopeRequest:
      type: object
      required: [roleId, scopeType]
      properties:
        operatorId:
          type: integer
          format: int64
          nullable: true
          description: 操作人用户ID（初始化阶段可为空）
          example: 1
        roleId:
          type: integer
          format: int64
          description: 角色ID
          example: 3
        scopeType:
          type: string
          description: ALL/SELF/DEPT/DEPT_AND_CHILD/CUSTOM
          example: DEPT_AND_CHILD
        deptId:
          type: integer
          format: int64
          nullable: true
          description: 部门ID（scopeType=CUSTOM时必填）
          example: 2
...
    RbacUserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID
          example: 10
        username:
          type: string
          description: 用户名
          example: zhangsan
        deptId:
          type: integer
          format: int64
          nullable: true
          description: 部门ID
          example: 2
        status:
          type: integer
          description: 状态（0禁用/1启用）
          example: 1
    SysRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 角色ID
          example: 3
        roleCode:
          type: string
          description: 角色编码
          example: EMPLOYEE
        roleName:
          type: string
          description: 角色名称
          example: 普通员工
        status:
          type: integer
          description: 状态（0禁用/1启用）
          example: 1
    SysDept:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 部门ID
          example: 2
        parentId:
          type: integer
          format: int64
          nullable: true
          description: 上级部门ID
          example: 1
        deptName:
          type: string
          description: 部门名称
          example: 财务部
    SysPost:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 岗位ID
          example: 5
        postCode:
          type: string
          description: 岗位编码
          example: FINANCE_MANAGER
        postName:
          type: string
          description: 岗位名称
          example: 财务经理